+++ lib/terradactyl/cli.rb
@@ -68,9 +68,15 @@ module Terradactyl
         Terradactyl::Terraform::VersionManager.latest
       end
 
-      def upgrade_stack(name)
-        @stack ||= Stack.new(name)
-        print_warning "Upgrading: #{@stack.name}"
+      # get name of base folder for printing logs
+      def base_folder_name(base_folder)
+        base_folder || config.misc.base_folder_name || config.base_folder
+      end
+
+      # rubocop:disable Metrics/AbcSize
+      def upgrade_stack(name, base_override = nil)
+        @stack ||= Stack.new(name, base_override)
+        print_warning "Upgrading in #{config.base_folder}: #{@stack.name}"
         if @stack.upgrade.zero?
           print_ok "Upgraded: #{@stack.name}"
         else
@@ -95,10 +101,10 @@ module Terradactyl
       puts config.to_h.to_yaml
     end
 
-    desc 'stacks', 'List the stacks'
-    def stacks
+    desc 'stacks [BASE_FOLDER]', 'List the stacks, with optional base folder override'
+    def stacks(base_override = nil)
       print_ok 'Stacks:'
-      Stacks.load.each do |name|
+      Stacks.load(base_override: base_override).each do |name|
         print_dot name.to_s
       end
     end
@@ -114,49 +120,59 @@ module Terradactyl
     # * Some are useful only in pipelines. These are hidden.
     #################################################################
 
-    desc 'planpr', 'Plan stacks against origin/HEAD (used for PRs)', hide: true
-    def planpr
-      print_header 'SmartPlanning PR ...'
-      stacks = Stacks.load(filter: StacksPlanFilterGitDiffOriginBranch.new)
+    desc 'planpr [BASE_FOLDER]',
+         'Plan stacks against origin/HEAD (used for PRs), with optional base folder override',
+         hide: true
+    def planpr(base_override = nil)
+      print_header "SmartPlanning PR in #{base_folder_name(base_override)} ..."
+      stacks = Stacks.load(
+        filter: StacksPlanFilterGitDiffOriginBranch.new,
+        base_override: base_override
+      )
       validate_planpr(stacks).each do |name|
-        clean(name)
-        init(name)
-        plan(name)
+        clean(name, base_override)
+        init(name, base_override)
+        plan(name, base_override)
         @stack = nil
       end
     end
 
-    desc 'smartplan', 'Plan any stacks that differ from Git HEAD'
-    def smartplan
-      print_header 'SmartPlanning Stacks ...'
-      stacks = Stacks.load(filter: StacksPlanFilterGitDiffHead.new)
+    desc 'smartplan [BASE_FOLDER]',
+         'Plan any stacks that differ from Git HEAD, with optional base folder override'
+    def smartplan(base_override = nil)
+      print_header "SmartPlanning Stacks in #{base_folder_name(base_override)} ..."
+      stacks = Stacks.load(filter: StacksPlanFilterGitDiffHead.new, base_override: base_override)
       validate_smartplan(stacks).each do |name|
-        clean(name)
-        init(name)
-        plan(name)
+        clean(name, base_override)
+        init(name, base_override)
+        plan(name, base_override)
         @stack = nil
       end
     end
 
-    desc 'smartapply', 'Apply any stacks that contain plan files', hide: true
-    def smartapply
-      print_header 'SmartApplying Stacks ...'
-      stacks = Stacks.load(filter: StacksApplyFilterPrePlanned.new)
+    desc 'smartapply',
+         'Apply any stacks that contain plan files, with optional base folder override',
+         hide: true
+    def smartapply(base_override = nil)
+      print_header "SmartApplying Stacks in #{base_folder_name(base_override)} ..."
+      stacks = Stacks.load(filter: StacksApplyFilterPrePlanned.new, base_override: base_override)
       print_warning 'No stacks contain plan files ...' unless stacks.any?
       stacks.each do |name|
-        apply(name)
+        apply(name, base_override)
         @stack = nil
       end
       print_message "Total Stacks Modified: #{stacks.size}"
     end
 
-    desc 'smartrefresh', 'Refresh any stacks that contain plan files', hide: true
-    def smartrefresh
-      print_header 'SmartRefreshing Stacks ...'
-      stacks = Stacks.load(filter: StacksApplyFilterPrePlanned.new)
+    desc 'smartrefresh [BASE_FOLDER]',
+         'Refresh any stacks that contain plan files, with optional base folder override',
+         hide: true
+    def smartrefresh(base_override = nil)
+      print_header "SmartRefreshing Stacks in #{base_folder_name(base_override)} ..."
+      stacks = Stacks.load(filter: StacksApplyFilterPrePlanned.new, base_override: base_override)
       print_warning 'No stacks contain plan files ...' unless stacks.any?
       stacks.each do |name|
-        refresh(name)
+        refresh(name, base_override)
         @stack = nil
       end
       print_message "Total Stacks Refreshed: #{stacks.size}"
@@ -168,56 +184,57 @@ module Terradactyl
     # the `quickplan` task is an exception to this rule.
     #################################################################
 
-    desc 'upgrade NAME', 'Cleans, inits, upgrades and formats an individual stack, by name'
-    def upgrade(name)
-      clean(name)
-      init(name, backend: false)
-      upgrade_stack(name)
-      fmt(name)
+    desc 'upgrade NAME [BASE_FOLDER]',
+         'Cleans, inits, upgrades and fmts a given stack, by name & optional base folder override'
+    def upgrade(name, base_override = nil)
+      clean(name, base_override)
+      init(name, base_override, backend: false)
+      upgrade_stack(name, base_override)
+      fmt(name, base_override)
     end
 
-    desc 'quickplan NAME', 'Clean, init and plan a stack, by name'
-    def quickplan(name)
+    desc 'quickplan NAME [BASE_FOLDER]',
+         'Clean, init and plan a stack, by name & optional base folder override'
+    def quickplan(name, base_override = nil)
       print_header "Quick planning #{name} ..."
-      clean(name)
-      init(name)
-      plan(name)
+      clean(name, base_override)
+      init(name, base_override)
+      plan(name, base_override)
     end
 
-    desc 'clean-all', 'Clean all stacks'
-    def clean_all
-      print_header 'Cleaning ALL Stacks ...'
-      Stacks.load.each do |name|
-        clean(name)
+    desc 'clean-all [BASE_FOLDER]', 'Clean all stacks, by optional base folder override'
+    def clean_all(base_override = nil)
+      print_header "Cleaning ALL Stacks in #{base_folder_name(base_override)} ..."
+      Stacks.load(base_override: base_override).each do |name|
+        clean(name, base_override)
         @stack = nil
       end
     end
 
-    desc 'plan-all', 'Plan all stacks'
-    def plan_all
-      print_header 'Planning ALL Stacks ...'
-      Stacks.load.each do |name|
+    desc 'plan-all [BASE_FOLDER]', 'Plan all stacks, by optional base folder override'
+    def plan_all(base_override = nil)
+      print_header "Planning ALL Stacks in #{base_folder_name(base_override)} ..."
+      Stacks.load(base_override: base_override).each do |name|
         catch(:error) do
-          clean(name)
-          init(name)
-          plan(name)
+          clean(name, base_override)
+          init(name, base_override)
+          plan(name, base_override)
         end
         @stack = nil
       end
     end
 
-    desc 'audit-all', 'Audit all stacks'
+    desc 'audit-all [BASE_FOLDER]', 'Audit all stacks, by optional base folder override'
     options report: :optional
     method_option :report, type: :boolean
-    # rubocop:disable Metrics/AbcSize
-    def audit_all
+    def audit_all(base_override = nil)
       report = { start: Time.now.to_json }
-      print_header 'Auditing ALL Stacks ...'
-      Stacks.load.each do |name|
+      print_header "Auditing ALL Stacks in #{base_folder_name(base_override)} ..."
+      Stacks.load(base_override: base_override).each do |name|
         catch(:error) do
-          clean(name)
-          init(name)
-          audit(name)
+          clean(name, base_override)
+          init(name, base_override)
+          audit(name, base_override)
         end
         @stack = nil
       end
@@ -229,14 +246,14 @@ module Terradactyl
     end
     # rubocop:enable Metrics/AbcSize
 
-    desc 'validate-all', 'Validate all stacks'
-    def validate_all
-      print_header 'Validating ALL Stacks ...'
-      Stacks.load.each do |name|
+    desc 'validate-all [BASE_FOLDER]', 'Validate all stacks, by optional base folder override'
+    def validate_all(base_override = nil)
+      print_header "Validating ALL Stacks in #{base_folder_name(base_override)} ..."
+      Stacks.load(base_override: base_override).each do |name|
         catch(:error) do
-          clean(name)
-          init(name)
-          validate(name)
+          clean(name, base_override)
+          init(name, base_override)
+          validate(name, base_override)
         end
         @stack = nil
       end
@@ -247,9 +264,10 @@ module Terradactyl
     # * These tasks are used regularly against stacks, by name.
     #################################################################
 
-    desc 'lint NAME', 'Lint an individual stack, by name'
-    def lint(name)
-      @stack ||= Stack.new(name)
+    desc 'lint NAME [BASE_FOLDER]',
+         'Lint an individual stack, by name & optional base folder override'
+    def lint(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
       print_ok "Linting: #{@stack.name}"
       if @stack.lint.zero?
         print_ok "Formatting OK: #{@stack.name}"
@@ -259,9 +277,10 @@ module Terradactyl
       end
     end
 
-    desc 'fmt NAME', 'Format an individual stack, by name'
-    def fmt(name)
-      @stack ||= Stack.new(name)
+    desc 'fmt NAME [BASE_FOLDER]',
+         'Format an individual stack, by name & optional base folder override'
+    def fmt(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
       print_warning "Formatting: #{@stack.name}"
       if @stack.fmt.zero?
         print_ok "Formatted: #{@stack.name}"
@@ -271,12 +290,14 @@ module Terradactyl
       end
     end
 
-    desc 'init NAME', 'Init an individual stack, by name'
-    def init(name, backend: true)
-      @stack ||= Stack.new(name)
+    desc 'init NAME [BASE_FOLDER]',
+         'Init an individual stack, by name & optional base folder override'
+    # rubocop:disable Metrics/AbcSize
+    def init(name, base_override = nil, backend: true)
+      @stack ||= Stack.new(name, base_override)
       @stack.config.terraform.init.backend = backend
 
-      print_ok "Initializing: #{@stack.name}"
+      print_ok "Initializing in #{base_folder_name(base_override)}: #{@stack.name}"
       if @stack.init.zero?
         print_ok "Initialized: #{@stack.name}"
       else
@@ -286,11 +307,11 @@ module Terradactyl
       end
     end
 
-    desc 'plan NAME', 'Plan an individual stack, by name'
-    # rubocop:disable Metrics/AbcSize
-    def plan(name)
-      @stack ||= Stack.new(name)
-      print_ok "Planning: #{@stack.name}"
+    desc 'plan NAME [BASE_FOLDER]',
+         'Plan an individual stack, by name & optional base folder override'
+    def plan(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
+      print_ok "Planning in #{base_folder_name(base_override)}: #{@stack.name}"
       case @stack.plan
       when 0
         print_ok "No changes: #{@stack.name}"
@@ -309,19 +330,21 @@ module Terradactyl
     end
     # rubocop:enable Metrics/AbcSize
 
-    desc 'audit NAME', 'Audit an individual stack, by name'
-    def audit(name)
-      plan(name)
+    desc 'audit NAME [BASE_FOLDER]',
+         'Audit an individual stack, by name & optional base folder override'
+    def audit(name, base_override = nil)
+      plan(name, base_override)
       if (@stack = Stacks.dirty?(name))
         Stacks.error!(@stack)
         print_crit "Dirty stack: #{@stack.name}"
       end
     end
 
-    desc 'validate NAME', 'Validate an individual stack, by name'
-    def validate(name)
-      @stack ||= Stack.new(name)
-      print_ok "Validating: #{@stack.name}"
+    desc 'validate NAME [BASE_FOLDER]',
+         'Validate an individual stack, by name & optional base folder override'
+    def validate(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
+      print_ok "Validating in #{base_folder_name(base_override)}: #{@stack.name}"
       if @stack.validate.zero?
         print_ok "Validated: #{@stack.name}"
       else
@@ -331,18 +354,20 @@ module Terradactyl
       end
     end
 
-    desc 'clean NAME', 'Clean an individual stack, by name'
-    def clean(name)
-      @stack ||= Stack.new(name)
-      print_warning "Cleaning: #{@stack.name}"
+    desc 'clean NAME [BASE_FOLDER]',
+         'Clean an individual stack, by name & optional base folder override'
+    def clean(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
+      print_warning "Cleaning in #{config.base_folder}: #{@stack.name}"
       @stack.clean
       print_ok "Cleaned: #{@stack.name}"
     end
 
-    desc 'apply NAME', 'Apply an individual stack, by name'
-    def apply(name)
-      @stack ||= Stack.new(name)
-      print_warning "Applying: #{@stack.name}"
+    desc 'apply NAME [BASE_FOLDER]',
+         'Apply an individual stack, by name & optional base folder override'
+    def apply(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
+      print_warning "Applying in #{base_folder_name(base_override)}: #{@stack.name}"
       if @stack.apply.zero?
         print_ok "Applied: #{@stack.name}"
       else
@@ -351,10 +376,11 @@ module Terradactyl
       end
     end
 
-    desc 'refresh NAME', 'Refresh state on an individual stack, by name'
-    def refresh(name)
-      @stack ||= Stack.new(name)
-      print_crit "Refreshing: #{@stack.name}"
+    desc 'refresh NAME [BASE_FOLDER]',
+         'Refresh state on an individual stack, by name & optional base folder override'
+    def refresh(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
+      print_crit "Refreshing in #{base_folder_name(base_override)}: #{@stack.name}"
       if @stack.refresh.zero?
         print_warning "Refreshed: #{@stack.name}"
       else
@@ -363,10 +389,11 @@ module Terradactyl
       end
     end
 
-    desc 'destroy NAME', 'Destroy an individual stack, by name'
-    def destroy(name)
-      @stack ||= Stack.new(name)
-      print_crit "Destroying: #{@stack.name}"
+    desc 'destroy NAME [BASE_FOLDER]',
+         'Destroy an individual stack, by name & optional base folder override'
+    def destroy(name, base_override = nil)
+      @stack ||= Stack.new(name, base_override)
+      print_crit "Destroying in #{base_folder_name(base_override)}: #{@stack.name}"
       if @stack.destroy.zero?
         print_warning "Destroyed: #{@stack.name}"
       else